[Refactor]: Split asselist components into their own
