+++ cardstack/src/components/AssetList/AssetList.tsx
@@ -12,25 +12,20 @@ import {
   ActivityIndicator,
   StyleSheet,
 } from 'react-native';
-import { BackgroundColorProps, ColorProps } from '@shopify/restyle';
 import { getConstantByNetwork } from '@cardstack/cardpay-sdk';
 import { useRoute } from '@react-navigation/core';
 import AddFundsInterstitial from '../../../../src/components/AddFundsInterstitial';
-import { Theme } from '../../theme';
-import { PinnedHiddenSectionMenu } from '../PinnedHiddenSectionMenu';
-import { AssetFooter } from './AssetFooter';
-import { AssetListLoading } from './AssetListLoading';
+import { AssetListLoading } from './components/AssetListLoading';
+import { AssetListProps, AssetListRouteType } from './types';
+import { strings } from './strings';
+import AssetSectionHeader from './components/AssetSectionHeader';
+import { PinHideOptionsFooter } from '@cardstack/components/PinnedHiddenSection';
 import {
   DiscordPromoBanner,
   useDiscordPromoBanner,
 } from '@cardstack/components/DiscordPromoBanner';
-import {
-  Button,
-  Container,
-  ListEmptyComponent,
-  Text,
-} from '@cardstack/components';
-import { isLayer1, dateFormatter, Device } from '@cardstack/utils';
+import { Container, Text } from '@cardstack/components';
+import { isLayer1 } from '@cardstack/utils';
 import {
   PinnedHiddenSectionOption,
   useAccountProfile,
@@ -41,75 +36,6 @@ import showWalletErrorAlert from '@rainbow-me/helpers/support';
 import { useNavigation } from '@rainbow-me/navigation';
 import Routes from '@rainbow-me/routes';
 import logger from 'logger';
-import { Network } from '@rainbow-me/helpers/networkTypes';
-import {
-  AssetWithNativeType,
-  CollectibleType,
-  DepotType,
-  MerchantSafeType,
-  PrepaidCardType,
-} from '@cardstack/types';
-
-interface HeaderItem {
-  title: string;
-  count?: number;
-  showContextMenu?: boolean;
-  total?: string;
-  type?: PinnedHiddenSectionOption;
-}
-
-export type AssetListSectionItem<ComponentProps> = {
-  Component: (
-    props: ComponentProps & {
-      networkName: string;
-      nativeCurrency: string;
-      currencyConversionRates: {
-        [key: string]: number;
-      };
-    }
-  ) => JSX.Element | null;
-  header: HeaderItem;
-  data: ComponentProps[];
-  timestamp?: string;
-};
-
-type SectionType =
-  | PrepaidCardType
-  | DepotType
-  | MerchantSafeType
-  | AssetWithNativeType
-  | CollectibleType;
-
-interface AssetListProps
-  extends BackgroundColorProps<Theme>,
-    ColorProps<Theme> {
-  isEmpty: boolean;
-  loading: boolean;
-  network: Network;
-  nativeCurrency: string;
-  currencyConversionRates: {
-    [key: string]: number;
-  };
-  sections: AssetListSectionItem<SectionType>[];
-  headerPaddingVertical?: number;
-  headerPaddingHorizontal?: number;
-  refetchSafes: () => void;
-  isFetchingSafes: boolean;
-}
-
-interface RouteType {
-  params: {
-    scrollToPrepaidCardsSection?: boolean;
-    forceRefreshOnce?: boolean;
-  };
-  key: string;
-  name: string;
-}
-
-// We need to pass this prop if the section to scrollTo is not on viewport
-const onScrollToIndexFailed = () => {
-  logger.log('onScrollToIndexFailed');
-};
 
 const styles = StyleSheet.create({
   contentContainer: {
@@ -117,17 +43,9 @@ const styles = StyleSheet.create({
   },
 });
 
-const strings = {
-  lastUpdatedAt: (timestamp: string) =>
-    `Last updated at: ${dateFormatter(
-      parseFloat(timestamp),
-      'MM/dd/yy',
-      'h:mm a',
-      ', '
-    )}`,
-  newCardLabel: 'New Card',
-  buyCardLabel: 'Buy Prepaid Card',
-  emptyCardMessage: `You don't own any\nPrepaid Cards yet`,
+// We need to pass this prop if the section to scrollTo is not on viewport
+const onScrollToIndexFailed = () => {
+  logger.log('onScrollToIndexFailed');
 };
 
 export const AssetList = (props: AssetListProps) => {
@@ -136,7 +54,7 @@ export const AssetList = (props: AssetListProps) => {
   const [refreshing, setRefreshing] = useState(false);
   const { accountAddress } = useAccountProfile();
   const { navigate, setParams } = useNavigation();
-  const { params } = useRoute<RouteType>();
+  const { params } = useRoute<AssetListRouteType>();
 
   const { isDamaged } = useWallets();
 
@@ -213,14 +131,6 @@ export const AssetList = (props: AssetListProps) => {
     navigate(Routes.BUY_PREPAID_CARD);
   }, [isDamaged, navigate]);
 
-  const prevAccount = useRef(null);
-
-  useEffect(() => {
-    if (accountAddress) {
-      prevAccount.current = accountAddress;
-    }
-  }, [accountAddress]);
-
   const renderSectionHeader = useCallback(
     ({ section }) => (
       <AssetSectionHeader
@@ -250,6 +160,14 @@ export const AssetList = (props: AssetListProps) => {
       ) : null,
     [isFetchingSafes]
   );
+
+  const prevAccount = useRef(null);
+
+  useEffect(() => {
+    if (accountAddress) {
+      prevAccount.current = accountAddress;
+    }
+  }, [accountAddress]);
 
   // Account was switched so show loading skeleton
   const isLoadingSafesDiffAccount = useMemo(
@@ -291,85 +209,7 @@ export const AssetList = (props: AssetListProps) => {
         renderSectionHeader={renderSectionHeader}
         contentContainerStyle={styles.contentContainer}
       />
-      <AssetFooter />
-    </>
-  );
-};
-
-interface AssetSectionProps {
-  section: AssetListSectionItem<SectionType>;
-  onBuyCardPress: () => void;
-}
-
-const AssetSectionHeader = ({ section, onBuyCardPress }: AssetSectionProps) => {
-  const {
-    header: { type, title, count, showContextMenu, total },
-    data,
-  } = section;
-
-  const isEmptyPrepaidCard =
-    type === PinnedHiddenSectionOption.PREPAID_CARDS && data.length === 0;
-
-  const renderNewCardButton = useMemo(
-    () =>
-      Device.isIOS && (
-        <Button variant="tinyOpacity" onPress={onBuyCardPress}>
-          {strings.newCardLabel}
-        </Button>
-      ),
-    [onBuyCardPress]
-  );
-
-  return (
-    <>
-      <Container
-        alignItems="center"
-        flexDirection="row"
-        justifyContent="space-between"
-        padding={4}
-        backgroundColor="backgroundBlue"
-      >
-        <Container flexDirection="row">
-          <Text color="white" size="medium">
-            {title}
-          </Text>
-          {count ? (
-            <Text color="tealDark" size="medium" marginLeft={2}>
-              {count}
-            </Text>
-          ) : null}
-        </Container>
-        <Container alignItems="center" flexDirection="row">
-          {total ? (
-            <Text
-              color="tealDark"
-              size="body"
-              weight="extraBold"
-              marginRight={showContextMenu ? 3 : 0}
-            >
-              {total}
-            </Text>
-          ) : null}
-          {isEmptyPrepaidCard
-            ? renderNewCardButton
-            : showContextMenu && <PinnedHiddenSectionMenu type={type} />}
-        </Container>
-      </Container>
-      {isEmptyPrepaidCard && (
-        <Container marginHorizontal={4} alignItems="center">
-          <ListEmptyComponent
-            text={strings.emptyCardMessage}
-            width="100%"
-            hasRoundBox
-            textColor="blueText"
-          />
-          {Device.isIOS && (
-            <Button onPress={onBuyCardPress} marginTop={4}>
-              {strings.buyCardLabel}
-            </Button>
-          )}
-        </Container>
-      )}
+      <PinHideOptionsFooter />
     </>
   );
 };
