+++ cardstack/src/components/AssetList/index.ts
@@ -1,2 +1,1 @@
 export * from './AssetList';
-export * from './AssetFooter';
