+++ cardstack/src/components/AssetList/strings.ts
@@ -1,0 +1,14 @@
+import { dateFormatter } from '@cardstack/utils';
+
+export const strings = {
+  lastUpdatedAt: (timestamp: string) =>
+    `Last updated at: ${dateFormatter(
+      parseFloat(timestamp),
+      'MM/dd/yy',
+      'h:mm a',
+      ', '
+    )}`,
+  newCardLabel: 'New Card',
+  buyCardLabel: 'Buy Prepaid Card',
+  emptyCardMessage: `You don't own any\nPrepaid Cards yet`,
+};
