+++ cardstack/src/components/PinnedHiddenSection/PinnedHiddenSectionMenu.tsx
