+++ cardstack/src/components/PinnedHiddenSection/index.ts
@@ -1,1 +1,2 @@
 export { default as PinnedHiddenSectionMenu } from './PinnedHiddenSectionMenu';
+export { default as PinHideOptionsFooter } from './PinHideOptionsFooter';
